Frontend to Backend First Implementation
These files are a super basic first implementation of connecting the front end HTML to the backend PHP (by extension, the database file as well). In this case, the signup.html and signup.php is updated. If you want to test these out, change the file names to signup.html and signup.php. (I don't want to cause confusion to the earlier files committed.)